add brand exceptions

+++ StockSuiteMono/src/main/java/org/crishof/stocksuitemono/service/BrandServiceImpl.java
@@ -1,8 +1,11 @@
 package org.crishof.stocksuitemono.service;
 
+import org.crishof.stocksuitemono.exception.duplicated.DuplicateNameException;
 import org.crishof.stocksuitemono.exception.notFound.BrandNotFoundException;
 import org.crishof.stocksuitemono.model.Brand;
+import org.crishof.stocksuitemono.model.Product;
 import org.crishof.stocksuitemono.repository.BrandRepository;
+import org.crishof.stocksuitemono.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +17,8 @@ public class BrandServiceImpl implements BrandService {
 
     @Autowired
     BrandRepository brandRepository;
+    @Autowired
+    ProductRepository productRepository;
 
     @Override
     public List<Brand> getAll() {
@@ -28,6 +33,10 @@ public class BrandServiceImpl implements BrandService {
 
     @Override
     public Brand save(Brand brand) {
+
+        if (brandRepository.findByNameIgnoreCase(brand.getName()).isPresent()) {
+            throw new DuplicateNameException("Brand with name " + brand.getName() + " already exist");
+        }
         return brandRepository.save(brand);
     }
 
@@ -45,6 +54,12 @@ public class BrandServiceImpl implements BrandService {
 
     @Override
     public void deleteById(UUID id) {
+
+        Brand brand = brandRepository.findById(id).orElseThrow(() -> new BrandNotFoundException(id));
+        List<Product> products = productRepository.findAllByBrandName(brand.getName());
+        if (!products.isEmpty()) {
+            throw new IllegalStateException("Cannot delete brand with asociated products");
+        }
         brandRepository.deleteById(id);
     }
